♻️ refactor:index.ts 수정
