--- nova-app/src/features/index.ts
@@ -1,1 +1,0 @@
-// init
