+++ nova-app/src/features/trend/ui/TrendChart.tsx
@@ -4,7 +4,7 @@ import { Line } from 'react-chartjs-2';
 import '../lib/chart';
 import { trendChartData } from '@/features/trend/mock/lineChartData';
 import createTrendOptions from '@/features/trend/lib/lineOption';
-import TrendLegend, { LegendDataset } from '@/features/trend/ui/ChartLegend';
+import TrendLegend, { LegendDataset } from '@/features/trend/ui/TrendLegend';
 const TrendChart = () => {
   return (
     <div className='bg-static rounded-2xl'>
