+++ nova-app/src/features/trend/ui/TrendLegend.tsx
