+++ nova-app/src/features/trend/ui/index.ts
@@ -1,0 +1,5 @@
+export { default as BarChart } from './BarChart';
+export { default as CategoriesKeyword } from './CategoriesKeyword';
+export { default as TrendChart } from './TrendChart';
+export { default as KeywordTop } from './KeywordTop';
+export { default as TrendLegend } from './TrendLegend';
